added size to p-space and changed size to double

+++ Garage3.0/Models/VehicleType.cs
@@ -6,7 +6,6 @@ namespace Garage3.Models
     {
         [Key]
         public string Type { get; set; }
-        public int Size { get; set; }
-
+        public double Size { get; set; }
     }
 }
